requirements.txt
apturl==0.5.2
async-generator==1.10
attrs==21.4.0
bcrypt==3.1.7
beautifulsoup4==4.11.1
blinker==1.4
Brlapi==0.7.0
CacheControl==0.12.10
cachetools==4.2.4
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
colorama==0.4.3
colour==0.1.5
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
daemonize==2.4.7
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
fasteners==0.14.1
firebase-admin==5.2.0
future==0.18.2
google-api-core==1.30.0
google-api-python-client==2.8.0
google-auth==1.30.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
grpcio==1.44.0
h11==0.13.0
httplib2==0.20.4
huepy==1.2.1
idna==2.8
imageio==2.16.2
imageio-ffmpeg==0.4.6
iniconfig==1.1.1
instabot==0.5.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mock==4.0.3
monotonic==1.5
moviepy==1.0.3
msgpack==1.0.3
netifaces==0.10.4
notify2==0.3
numpy==1.22.3
oauth2client==4.1.3
oauthlib==3.1.0
olefile==0.46
outcome==1.1.0
packaging==21.3
paramiko==2.6.0
pbr==5.8.1
pexpect==4.6.0
Pillow==9.1.0
pluggy==1.0.0
proglog==0.1.9
proto-plus==1.20.3
protobuf==3.20.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PyQt5==5.14.1
PyQtWebEngine==5.14.0
pyRFC3339==1.1
PySocks==1.7.1
pytest==7.1.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
responses==0.20.0
rsa==4.8
schedule==1.1.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
selenium==4.1.3
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
systemd-python==234
testresources==2.0.1
tomli==2.0.1
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.9
wadllib==1.3.3
wsproto==1.1.0
xkit==0.0.0
